+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 - PIM-10967: Fix inconsistency on DQI completeness recommendation
 - PIM-10639: Prevent users to change his password without providing its current password
 - PIM-10958: Fix attribute option position after clicking on "done"
+- PIM-10976: Fix variant product counter on Product Model Edit Form for variant products without identifier
 
 ## Improvements